requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
langchain==0.3.27
langchain-groq==0.3.8
langchain-openai==0.2.2
langchain-community==0.3.27
langchain-core==0.3.76
langchain-text-splitters==0.3.11
faiss-cpu==1.8.0
python-multipart==0.0.9
arxiv==2.1.3
wikipedia==1.4.0
pypdf